Warn users about cross-site problems
When using different top and second-level domains for GitHub Enterprise Server and ZenHub On-Prem, it can lead to issues in the browser since the traffic is cross-site. The main issue we've seen is a failure to load user avatars on the ZenHub Board in the web app when using Chrome.